backend/requirements.txt
# ============================================================
# DEPENDENCIAS DEL BACKEND (Django REST Framework)
# ============================================================
asgiref==3.10.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.3.1
django-filter==25.2
django-simple-history==3.7.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
pillow==12.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
reportlab==4.4.5
setuptools<81
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2


